ai-task-bot/requirements.txt
aiogram>=3.0.0
celery[redis]>=5.3.0
redis[hiredis]>=4.5.0
python-dotenv>=1.0.0
